Add files via upload
This marks the beginning of my first react project which is a To-Do List. This will comprise of the skills I have been learning and uploading for the past 2 week, such as states, conditional and/or ternary operators. This current upload displays the basis of the To-Do List; rendering lists to the UI.